moved test data to startup directory in imports directory
